created validation

+++ app/Http/Controllers/Country/CountryController.php
@@ -5,6 +5,9 @@ namespace App\Http\Controllers\country;
 use App\Http\Controllers\Controller;
 use App\Models\CountryModel;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
+
+use function PHPUnit\Framework\isNull;
 
 class CountryController extends Controller
 {
@@ -13,18 +16,38 @@ class CountryController extends Controller
     }
 
     public function countryByID($id){
-        return response()-> json(CountryModel::find($id), 200);
+        $country= CountryModel::find($id);
+        if(is_null($country)){
+            return response()-> json(["message"=>"Record not found"], 404);
+        }
+        return response()-> json($country, 200);
     }
     public function countrySave(Request $request){
+        $rules = [
+            'name' => 'required|min:3',
+            'iso' => 'required|min:2|max:2|'
+        ];
+        $validator = Validator::make($request->all(), $rules);
+        if($validator->fails()){
+            return response()->json($validator->errors(),400);
+        }
         $country = CountryModel::create($request->all());
         return response()->json($country,201);
     }
 
-    public function countryUpdate(Request $request, CountryModel $country){
+    public function countryUpdate(Request $request, $id){
+        $country = CountryModel::find($id);
+        if (is_null($country)){
+            return response()-> json(["message"=>"Record not found"], 404);
+        }
         $country-> update($request->all());
         return response()->json($country,200);
     }
-    public function countryDelete(Request $request, CountryModel $country){
+    public function countryDelete(Request $request, $id){
+        $country = CountryModel::find($id);
+        if(is_null($country)){
+            return response()-> json(["message"=>"Record not found"], 404);
+        }
         $country->delete();
         return response()->json(null, 204);
     }
